+++ src/main/java/mx/axxib/aforedigitalgt/eml/FopagosListOut.java
@@ -23,18 +23,18 @@ import lombok.NoArgsConstructor;
 //						
 //					@ColumnResult(name = "pCuenta", type = String.class),
 //					@ColumnResult(name = "pNombre", type = String.class),
-//					@ColumnResult(name = "pNoPoliza_Pag", type = Integer.class),
+//					@ColumnResult(name = "pNoPoliza_Pag", type = Long.class),
 //					@ColumnResult(name = "pReferencia", type = String.class),
 //					@ColumnResult(name = "pFechaMov_Pag", type = Date.class),
 //					@ColumnResult(name = "pFecReinv_Pag", type = Date.class),
 //					@ColumnResult(name = "pTelefono_Pag", type = String.class),
 //					@ColumnResult(name = "pBenef_Re", type = String.class),
-//					@ColumnResult(name = "pPorcent_Re", type = Integer.class),
+//					@ColumnResult(name = "pPorcent_Re", type = Double.class),
 //					@ColumnResult(name = "pEstatus_Re", type = String.class),
 //					@ColumnResult(name = "pPlaza_Re", type = String.class),
-//					@ColumnResult(name = "pMonBruto_Re", type = Integer.class),
-//					@ColumnResult(name = "pMontoIsr_Re", type = Integer.class),
-//					@ColumnResult(name = "pMontoNeto_Re", type = Integer.class),
+//					@ColumnResult(name = "pMonBruto_Re", type = Double.class),
+//					@ColumnResult(name = "pMontoIsr_Re", type = Double.class),
+//					@ColumnResult(name = "pMontoNeto_Re", type = Double.class),
 //					@ColumnResult(name = "pMensaje", type = String.class),
 //					@ColumnResult(name = "on_Estatus", type = Integer.class)
 //					
@@ -49,18 +49,18 @@ public class FopagosListOut {
 	
 	private String PCuenta;// OUT VARCHAR2,
 	private String PNombre;// OUT VARCHAR2,
-	private Integer PNoPoliza_Pag;// OUT NUMBER,
+	private Long PNoPoliza_Pag;// OUT NUMBER,
     private String PReferencia;// OUT VARCHAR2,
     private Date PFechaMov_Pag;// OUT DATE,
     private Date PFecReinv_Pag;// OUT DATE,
     private String PTelefono_Pag;// OUT VARCHAR2,
     private String PBenef_Re;// OUT VARCHAR2,
-    private Integer PPorcent_Re;// OUT NUMBER,
+    private Double PPorcent_Re;// OUT NUMBER,
     private String PEstatus_Re;// OUT VARCHAR2,
     private String PPlaza_Re;// OUT VARCHAR2,
-    private Integer PMonBruto_Re;// OUT NUMBER,
-    private Integer PMontoIsr_Re;// OUT NUMBER,
-    private Integer PMontoNeto_Re;// OUT NUMBER,
+    private Double PMonBruto_Re;// OUT NUMBER,
+    private Double PMontoIsr_Re;// OUT NUMBER,
+    private Double PMontoNeto_Re;// OUT NUMBER,
     private String PMensaje;// OUT VARCHAR2)
     private Integer On_Estatus;//OUT INTEGER
 }
